+++ tools/node/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Updated to support proto v0.55 release.
+- Added `pin_version` and `unpin_version` support, which maps to `package.json` `devEngines.packageManager`.
+
 #### 🐞 Fixes
 
 - Fixed some `package.json` version parsing issues.